Decoupling the blockchain logic from the network setup logic by using DI. Removing NodeNet from project. Renaming project to the "MicrocoinCore"
Such a separation will greatly simplify the testing and development of the project, the presence of a large number of bugs in the broadcast network system is currently noticed, therefore, in the near future, the replacement of this system with a plug-debugger is planned. After completing the blockchain core, I will return to NodeNet to rewrite it from scratch.